Fix the remaining issues with the linting and testing.
Update the GitHub Actions workflow to run the tests in a single workflow instead of two separate workflows.

The Makefile now does everything out of the container, so the GitHub Actions workflow can be simplified.

I've deleted the exception handler, as this gets in the way of viewing the error messages.