skip tests that require a live server
using 127.0.0.1, localhost or `socket.gethostname()` don't work. there's
always a connection refused error. it seems i'm not the only one with
this problem, and everoyne's solution is "just run mongo in another
container" which doesn't help at all. i can't find anything in the docs
on this, so we're taking the lazy approach and not wasting my time any
more. local tests will catch this if we really screw something up.
probably. hopefully.